wahuz/requirements.txt
streamlit==1.28.1
PyPDF2==3.0.1
numpy==1.24.3
google-generativeai==0.3.2
kafka-python==2.0.2
prometheus-client==0.19.0
requests==2.31.0
python-dotenv==1.0.0
mlflow==2.8.1
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
